Completed the Writing to the user table
The vibe backend can now read the transaction table and send this
information to the corresponding fields in the user table. This info is
readily available for whenever the dashboard is setup. The sql tables
have been changed due to integer overload problems and some other
considerations. They will have to be changed in all the environments to
match my localhost