News Feed API 1.1
Big overhaul to the Newsfeed API! Get standard newsfeed in ['friend']
field now, not ['data']. ['community'] containing the last 50 vibes from
the all the communities the user is in.

Note**: Check to make sure Limit 200 has uniform balance. If not the
order by descending step may have to be done before the 3rd join