newsfeed and newsfeed element //
Much cleaner design, but small step back as a result. Instead of doing
a fancy comparison method on the newsfeed element side, I simply do all
of the appending on the front-end dynamically and only refresh when the
user hits refresh. In the near future, I can build functionality that
triggers small banner if new elements are available like they do in
LinkedIn mobile.
— Noah